refactor: organizes folders and files names
