+++ src/components/icons/FavoriteIcon.tsx
