+++ src/components/icons/NavigationMenuIcon.tsx
