+++ src/components/newCollection/index.tsx
