--- src/pages/sapatos/sandalias/components/breadCrumb/index.tsx
@@ -1,16 +1,0 @@
-import { NavLink } from '@/pages/home/components/NavLink';
-import { navLinks } from '../../utils/nav-links';
-
-export function BreadCrumb() {
-  return (
-    <section className="w-full flex items-center h-12">
-      <nav className="flex space-x-2 px-5">
-        {navLinks.map((link, index) => (
-          <NavLink key={index} href={link.href}>
-            {link.label}
-          </NavLink>
-        ))}
-      </nav>
-    </section>
-  );
-}
