--- src/pages/sapatos/sandalias/index.tsx
@@ -1,11 +1,0 @@
-import { BreadCrumb } from './components/breadCrumb';
-import { TypesSandalsCards } from './components/typesSandalsCards';
-
-export default function SandalsPDC() {
-  return (
-    <div>
-      <BreadCrumb />
-      <TypesSandalsCards />
-    </div>
-  );
-}
