--- src/pages/sapatos/sandalias/utils/nav-links.ts
@@ -1,7 +1,0 @@
-export const navLinks = [
-  { label: 'home', href: '/' },
-  { label: '|', href: '' },
-  { label: 'sapatos', href: '/sapatos' },
-  { label: '|', href: '' },
-  { label: 'sandálias', href: '/sapatos/sandalias' },
-];
