+++ src/pages/sapatos/sandals/index.tsx
@@ -1,0 +1,11 @@
+import { BreadCrumb } from '../../../components/breadCrumb';
+import { CategoriesCard } from '../../../components/categoriesCard';
+
+export default function SandalsPDC() {
+  return (
+    <div>
+      <BreadCrumb />
+      <CategoriesCard />
+    </div>
+  );
+}
