+++ src/types/product.ts
